requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==2.1.15
django-cors-headers==3.4.0
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
elementpath==2.2.2
idna==2.10
importlib-resources==5.1.4
lxml==4.6.3
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
pyOpenSSL==20.0.1
pysaml2==7.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
xmlschema==1.6.2
xmlsec==1.3.11
zipp==3.4.1
